- Moved config structure to seperate files - Bugfix
