+++ HABApp/rule/watched_item.py
@@ -17,8 +17,12 @@ class WatchedItem:
     def check(self, now):
         if self.is_canceled:
             return None
-
-        item = Items.get_item( self.name)
+        
+        try:
+            item = Items.get_item( self.name)
+        except KeyError:
+            return None
+        
         timestamp = item.last_change if self.__watch_only_changes else item.last_update
         duration = now - timestamp
         if duration < self.const:
