+++ HABApp/util/threshold.py
@@ -2,7 +2,7 @@
 class Threshold:
     def __init__(self, threshold1, threshold2):
         """This is a simple Schmitt Trigger implementation.
-        If the value is >= upper_threshold is_higher will return true.
+        If the value is > upper_threshold is_higher will return true.
         The return value will stay true until the value goes below the lower threshold.
 
         :param threshold1:
@@ -14,13 +14,20 @@ class Threshold:
         assert self.lower_threshold <= self.upper_threshold
 
         self.__threshold = self.upper_threshold
+    
+    def is_on(self, value):
+        self.check_value(value)
+        return self.__threshold == self.lower_threshold
+    
+    def is_off(self, value):
+        return not self.is_on(value)
 
     @property
     def current_threshold(self):
         return self.__threshold
 
-    def __check_value(self, value):
-        if value >= self.upper_threshold:
+    def check_value(self, value):
+        if value > self.upper_threshold:
             self.__threshold = self.lower_threshold
 
         if value < self.lower_threshold:
@@ -32,22 +39,22 @@ class Threshold:
         if not isinstance(other, (int, float)):
             return NotImplemented
 
-        return self.__check_value(other) < other
+        return self.check_value(other) < other
 
     def __le__(self, other):
         if not isinstance(other, (int, float)):
             return NotImplemented
 
-        return self.__check_value(other) <= other
+        return self.check_value(other) <= other
 
     def __ge__(self, other):
         if not isinstance(other, (int, float)):
             return NotImplemented
 
-        return self.__check_value(other) >= other
+        return self.check_value(other) >= other
 
     def __gt__(self, other):
         if not isinstance(other, (int, float)):
             return NotImplemented
 
-        return self.__check_value(other) > other
+        return self.check_value(other) > other
